Places warning of no options info to console.warn
In the usecase when we want to place the smartbanner script in some
site-wide bundle but display banner (and thus use meta info) only
on some subpages, the script was throwing error which could interfere with
execution of other scripts.
The warning message was placed in console.warn instead and function returns
false if no options (meta info) detected.